Moving new method to DispatchedConfigAsyncJobExecutionActor only
