+++ supportedBackends/sfs/src/main/scala/cromwell/backend/impl/sfs/config/ConfigConstants.scala
@@ -16,6 +16,7 @@ object ConfigConstants {
   val KillConfig = "kill"
   val KillDockerConfig = "kill-docker"
   val CheckAliveConfig = "check-alive"
+  val ExitCodeTimeoutConfig = "exit-code-timeout"
   val RuntimeAttributesConfig = "runtime-attributes"
   val JobIdRegexConfig = "job-id-regex"
   val RunInBackgroundConfig = "run-in-background"
